Created web app to display camera feed
Created an express server and a static router to serve a html file
which uses socket.io to open a socket to server and listen for ‘image’
events containing the png data buffer.

A blob object is used for holding the buffer and creating a URL, and an
image object is drawn to the canvas with the URL as source.